tool/cli/pyproject.toml
[project]
name = "msgpact_cli"
version = "0.1.5"
description = "A CLI tool for variouos local MsgPact operations."
requires-python = ">=3.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.1"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.0.0"
toml = "^0.10.2"
lxml = "^5.3.1"
pyyaml = "^6.0.2"
jinja2 = "*"
msgpack = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "*"

[tool.poetry.scripts]
msgpact = "msgpact_cli.cli:main"
